==> scripts/2_CARTS.Rmd <==
---
title: "CARTS"
output: html_document
---

Configuración inicial

```{r}

require(pacman)

p_load(
  tidyverse,
  tidymodels,
  ggplot2,
  plotly,
  leaflet,
  sf,
  visdat,
  geojsonsf,
  geojsonio,
  jsonlite,
  purrr,
  stringr,
  tm,
  stringi,
  skimr,
  tidyverse, 
  rpart, 
  caret ,  
  rpart.plot, 
  Metrics, 
  MLmetrics, 
  ipred,  
  ranger, 
  themis,
  yardstick
  )

raw_data_path <- './data'
db_train <- readRDS(file.path(raw_data_path, 'train.rds'))
db_test <- readRDS(file.path(raw_data_path, 'test.rds'))

# Verificando la estructura de la base de datos 
str(db_train)  
head(db_train)
summary(db_train)

```

# CARTs

## Modelo

Entrenando un arbol para regresión con la especificación

```{r}

#Partición para evaluar el modelo 
library(caret)
set.seed(123)

key <- createDataPartition(db_train$price_millions, p = 0.8, list = FALSE)

train_set <- db_train[key, ]
val_set <- db_train[-key, ]


```

```{r}
a 
```

```{r}
#Predicción con base de validación 
price_hat <- predict(modelo_cart_1, val_set)

# Añadir predicciones al conjunto de validación
val_set$price_hat <- price_hat

#Creando el data frame
resultados <- val_set %>%
  select(price) %>%
  mutate(price_hat = price_hat)

# Calcular MAE y RMSE
mae_result <- mae(resultados, truth = price, estimate = price_hat)
rmse_result <- rmse(resultados, truth = price, estimate = price_hat)

# Mostrar resultados
print(mae_result)
print(rmse_result)

#Preparamos los resultados
price_test <- predict(modelo_cart_1, db_test)

entrega_1 <- data.frame(
  property_id = db_test$property_id,
  price = as.vector(price_test) 
)

entrega_1 <- entrega_1 %>%
  mutate(price = round(price)) %>%
  distinct(property_id, .keep_all = TRUE)

# Guardar el CSV
write.csv(entrega_1, file = "submission_CARTs_cp_001.csv", row.names = FALSE)
```
