==> scripts/3_Random_forest.Rmd <==
---
title: "random_forest"
output: null
date: "2025-05-14"
---

```{r}
require(pacman)

p_load(
  tidyverse,
  caret,
  Metrics,
  ranger,
  spatialsample,
  tidymodels,
  sf,
  yardstick
)
```

```{r}
data <- '../data'
train <- read_rds(file.path(data, 'train.rds'))
test <- read_rds(file.path(data, 'test.rds'))
```

```{r}
train <- train %>% 
  select(-c(
    'city', 
    'month', 
    'surface_total', 
    'surface_covered', 
    'rooms', 
    'operation_type', 
    'title', 
    'description', 
    'geometry',
    'localidad', 
    "localidad_fct",
    'upl', 
    'description_clean',
    'dataset',
    'title_clean',
    'price_millions'
            )) %>% 
  as.data.frame()

set.seed(1905)
# Creación de particiones:
# - 80% para entrenamiento (bd_tr)
# - 20% para validación (bd_validation)
# Método createDataPartition de caret preserva la distribución de la variable objetivo
inTrain <- createDataPartition(
  y = train$price,     # Variable objetivo
  p = .80,             # Proporción para entrenamiento
  list = FALSE
)

# Filtrado de datos:
bd_tr <- train %>% filter(row_number() %in% inTrain)       # Conjunto de entrenamiento
bd_validation  <- train %>% filter(!row_number() %in% inTrain)
```
# Especificación del modelo con todas las variables excepto algunas
```{r}
exclusion_vars <- c("price", "property_id", "geometry", "lon_raw", "lat_raw")

predictors <- setdiff(names(bd_tr), exclusion_vars)

rhs <- paste(predictors, collapse = " + ")

especification_forest <- as.formula(paste("price ~", rhs))
```
# Ajuste del modelo Random Forest
```{r}
# --- CONFIGURACIÓN DEL MODELO RANDOM FOREST CON VALIDACIÓN CRUZADA ESPACIAL ---

# Parámetros de control:
# - method = "10 fold CV": Usa validación cruzada con 10 folds
bd_tr_sf <- st_as_sf(
  bd_tr,
  coords = c("lon_raw", "lat_raw"),
  crs = 4326
)

set.seed(2405)
block_folds <- spatial_block_cv(bd_tr_sf, v = 10)

receta <-  recipe(especification_forest, data = bd_tr_sf) %>% 
  step_naomit(all_predictors(), all_outcomes()) %>%
  step_dummy(all_nominal_predictors()) %>%
  step_zv(all_predictors())  

modelo_rf <- rand_forest(
  mode = "regression",
  trees = 1000,
  mtry = tune(),
  min_n = tune(),
  ) %>% 
  set_engine(
    "ranger",
    splitrule = "variance",
    importance = "permutation"
    )

grid <- expand.grid(
  mtry = c(9, 18, 29, 37),
  min_n = c(2, 6, 10, 14)
)

workflow <- workflow() %>%
  add_model(modelo_rf) %>%
  add_recipe(receta)


tree_grid <- tune_grid(
  workflow,
  resamples = block_folds,
  grid = grid,
  metrics = metric_set(yardstick::mae),
  control = control_grid(save_pred = TRUE)
)

collect_metrics(tree_grid)

best_params <- select_best(tree_grid, metric = "mae")

final_wf <- finalize_workflow(workflow, best_params)

final_fit <- fit(final_wf, data = bd_tr)
```
# Cálculo de MAE en set de validación
```{r}
bd_pred_cv <- bd_validation %>%
  mutate(price_hat = predict(final_fit, new_data = bd_validation)$.pred) %>% 
  select(
    'property_id',
    'price',
    'price_hat'
  ) %>% 
  as.data.frame()

#MAE
mae_result2 <- mean(abs(bd_pred_cv$price - bd_pred_cv$price_hat))
print(mae_result2)
```
# Cálculo de RMSE en set de validación
```{r}
rmse_result2 <- sqrt(mean((bd_pred_cv$price_hat - bd_pred_cv$price)^2))
print(rmse_result2)
```
# Generación de predicciones fuera de muestra
```{r}
test <- test %>% 
  select(-c(
    'city', 
    'month', 
    'surface_total', 
    'surface_covered', 
    'rooms', 
    'operation_type', 
    'title', 
    'description', 
    'geometry',
    'localidad', 
    "localidad_fct",
    'upl', 
    'description_clean',
    'dataset',
    'title_clean',
    'price_millions'
            )) %>%
  as.data.frame()

pred <- predict(final_fit, new_data = test)$.pred

test <- test %>% 
  select('property_id') %>% 
  mutate(price = pred)
```
```{r}
write_csv(test, 'RF_mtry_29_minnode_2_cv_10.csv')
```
